# Conflicts:
#	Examples/UIExplorer/UIExplorer.xcodeproj/project.pbxproj
#	Examples/UIExplorer/UIExplorerUnitTests/RCTImageLoaderTests.m
#	Examples/UIExplorer/js/ImageExample.js
#	Libraries/CameraRoll/RCTCameraRollManager.m
#	Libraries/Image/Image.ios.js
#	Libraries/Image/RCTImageEditingManager.m
#	Libraries/Image/RCTImageLoader.h
#	Libraries/Image/RCTImageLoader.m
#	Libraries/Image/RCTImageView.m
#	Libraries/Image/RCTImageViewManager.m
#	Libraries/Image/RCTShadowVirtualImage.m
#	React/Base/RCTImageSource.h
#	React/Base/RCTImageSource.m

+++ local-cli/wrong-react-native.js
@@ -9,11 +9,17 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
-console.error([
-  '\033[31mLooks like you installed react-native globally, maybe you meant react-native-cli?',
-  'To fix the issue, run:\033[0m',
-  'npm uninstall -g react-native',
-  'npm install -g react-native-cli'
-].join('\n'));
+var script = process.argv[1];
+var installedGlobally = script.indexOf('node_modules/.bin/react-native') === -1;
 
-process.exit(1);
+if (installedGlobally) {
+  console.error([
+    '\033[31mLooks like you installed react-native globally, maybe you meant react-native-cli?',
+    'To fix the issue, run:\033[0m',
+    'npm uninstall -g react-native',
+    'npm install -g react-native-cli'
+  ].join('\n'));
+  process.exit(1);
+} else {
+  require('./cli').run();
+}
